+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## Unreleased
+- Added Wikidot logout command
+
 ## ~~0.2.1: Beta 2 patch 1~~ 0.2.2: Beta 2 patch 2
 - Fixed error when trying to pull pages with parent page